delete coment

+++ src/coment/comment.servise.ts
@@ -28,12 +28,20 @@ export class Commentservise{
         throw new BadRequestException(error.message)
        }         
     }
-    async deletecoment(comentId,userId){
+    async deletecoment(comentId,postId,userId){
            try {
             const coment=await this.comentmodel.findOne({_id:comentId})
+            if(!coment){
+                throw new BadRequestException("coment dont found")
+            }
             if(coment.userId != userId){
                 throw new BadRequestException("action dont alloed")
             }
+            const post =await this.postModel.findOne({_id:postId})
+            console.log("aaaaaaaaa",post.coments,coment._id)
+            post.coments=post.coments.filter((el)=>String(el) != String(coment._id))
+            console.log("ssssss",post.coments)
+            await post.save()
             await this.comentmodel.findOneAndDelete({_id:coment._id})
             return {message:"coment deleted"}
            } catch (error) {
